Monitoring_Project/requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dnspython==2.4.2
Flask==3.0.0
Flask-MQTT==1.1.1
fonttools==4.47.2
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mongoengine==0.27.0
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.1
numpy==1.26.3
packaging==23.2
paho-mqtt==1.6.1
pillow==10.2.0
ply==3.11
pyasn1==0.5.1
pycryptodomex==3.20.0
pymongo==4.6.1
pyparsing==3.1.1
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.2
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
threadpoolctl==3.2.0
urllib3==2.1.0
Werkzeug==3.0.1
